[mod] move some settings to settings.py
